test(runtime): isolate throw exception chain regression
